+++ ErgoJr/main.py
@@ -26,9 +26,8 @@ SOFTWARE.
 
 import time, sys
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
-sys.path.append("./Atlas")
+sys.path.append("./atlastk")
+sys.path.append("../atlastk")
 
 import atlastk as Atlas