+++ Hello/Hello.py
@@ -24,8 +24,8 @@ SOFTWARE.
 
 import os, sys
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
+sys.path.append("./atlastk")
+sys.path.append("../atlastk")
 
 import atlastk as Atlas
 
@@ -47,15 +47,15 @@ body = """
 </div>
 """
 
-def acConnect(self, dom):
+def acConnect(dom):
 	dom.setLayout("", body )
 	dom.focus( "input")
 
-def acSubmit(self, dom):
+def acSubmit(dom):
 	dom.alert("Hello, " + dom.getContent("input") + "!")
 	dom.focus( "input")
 
-def acClear(self, dom):
+def acClear(dom):
 	if ( dom.confirm("Are you sure?" ) ):
 		dom.setContent("input", "" )
 	dom.focus( "input")
@@ -66,4 +66,4 @@ callbacks = {
 	"Clear": acClear,
 }
 
-Atlas.launch(callbacks, lambda: None, head)
+Atlas.launch(callbacks, None, head)