+++ RGB/main.py
@@ -24,9 +24,8 @@ SOFTWARE.
 
 import GPIOq, sys
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
-sys.path.append("./Atlas")
+sys.path.append("./atlastk")
+sys.path.append("../atlastk")
 
 import atlastk as Atlas