+++ ReversiIMG/main.py
@@ -32,8 +32,8 @@ import random
 import itertools
 import time
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
+sys.path.append("./atlastk")
+sys.path.append("../atlastk")
 
 import atlastk as Atlas
 
@@ -259,4 +259,4 @@ callbacks = {
     "New": acNew
 }
 
-Atlas.launch(callbacks, lambda: Reversi(), readAsset("Head.html"), "ReversiIMG")
+Atlas.launch(callbacks, Reversi, readAsset("Head.html"), "ReversiIMG")