+++ ReversiXSL/main.py
@@ -34,8 +34,8 @@ import random
 import itertools
 import time
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
+sys.path.append("./atlastk")
+sys.path.append("../atlastk")
 
 import atlastk as Atlas